+++ internal/central/adapter/adapter_multi_unit_test.go
@@ -13187,7 +13187,7 @@ func (f *configFakeOperations) GetParamset(_ context.Context, _ string, _ hmenum
 	return f.paramsetData, f.paramsetErr
 }
 
-func (f *configFakeOperations) PutParamset(_ context.Context, _ string, _ hmenum.ParamsetKey, _ map[string]any, _ hmenum.CommandRxMode) error {
+func (f *configFakeOperations) PutParamset(_ context.Context, _ string, _ hmenum.ParamsetKey, _ map[string]any, _ hmenum.CommandPriority, _ hmenum.CommandRxMode) error {
 	return f.putErr
 }